requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.9.0
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
billiard==4.2.1
CacheControl==0.14.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.3
dataclasses-json==0.6.7
distro==1.9.0
exa_py==1.5.1
exceptiongroup==1.2.2
faiss-cpu==1.9.0
fastapi==0.115.4
fastapi-cli==0.0.5
firebase-admin==6.5.0
flatbuffers==24.3.25
frozenlist==1.5.0
gast==0.6.0
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-generativeai==0.8.3
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.62.3
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.5.0
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
keras==2.15.0
kombu==5.4.2
langchain==0.3.6
langchain-community==0.3.4
langchain-core==0.3.14
langchain-google-genai==2.0.3
langchain-text-splitters==0.3.1
langsmith==0.1.138
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.0
mdurl==0.1.2
ml-dtypes==0.2.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.53.0
opt_einsum==3.4.0
orjson==3.10.10
packaging==24.1
pillow==11.0.0
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.2.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.3
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
tenacity==9.0.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.15.0
termcolor==2.5.0
tqdm==4.66.6
twilio==9.3.6
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
Werkzeug==3.0.6
wrapt==1.14.1
yarl==1.17.1
zipp==3.20.2
